Request handler trait (#33)
* Adds a trait for handling SP requests in the host.

We previously created a `Controller` with a request handler callable
directly, something that basically looked like `Fn(Request) ->
Response`. That was fine to start, but we need a way to include some
state in the request-handling logic. This moves towards that goal by
making a `RequestHandler` trait, with a single method for dispatching
requests from the SP. That allows us to refer to `&self` when handling
those.

* Some re-exports and helpers