Handle requests via a channel (#44)
Removes the async-trait for a request handler, and instead opts to take
a send-half of a channel at construction. Rather than calling the
request handler method directly, the controller now sends the request on
that channel along with a oneshot for the reponse. It awaits the
response on that oneshot; if it's Some(_), the message is serialized and
sent back to the peer over the UDP socket; if it's None, the message is
dropped. This approach reduces the complexity around ownership and
generally allows more freedom to run the controller and handler in
separate tasks.