Handles software control of power mode when setting module power (#55)
* Handles software control of power mode when setting module power

This updates the various subcommands for managing a module's power state
to handle the case where software override of power control is set. It
keeps in place the existing, direct-control subcommands for setting the
disposition of hardware signals, and changes the `power_mode` and
`set_power_mode` functions to check for software override when
operating.

Note that one goal here is always keep the hardware and software signals
in sync. That is, if we request low-power mode, we set both the hardware
`LPMode` pin and the right bit in the module memory map. The idea is
that if one toggles the software-override, the power mode itself will
not change.

* Review feedback

- Print formatting
- Clearer filter/map
- Make `PowerControl` into an enum rather than a struct
- Handle modules which need power, but already have reset deasserted,
  when powering them into low-power mode. We emit a warning, assert
  reset, and then do the normal process of enabling power and
  deasserting reset (with wait times for each specified by SFF-8679).

* I take it back

* Assert LPMode hardware before deasserting ResetL